Finance-assistant-swarm-agent/pyproject.toml
[project]
name = "08-finance-assistant-swarm-agent"
version = "0.1.0"
description = "Finance-Assistant Swarm Agent Collaboration is a modular, multi-agent system designed to autonomously generate comprehensive equity research reports from a single stock query."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-requests-auth>=0.4.3",
    "boto3>=1.38.36",
    "frozendict>=2.4.6",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "requests>=2.32.4",
    "strands-agents>=0.1.7",
    "strands-agents-tools>=0.1.5",
    "textblob>=0.19.0",
    "finnhub-python>=2.4.18",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "PyYAML>=6.0.0",
    "python-dateutil>=2.8.0",
    "coloredlogs>=15.0.0", 
    "botocore>=1.31.0",
]
